delegate-event-to-control-behind-mask
When the drop down is opened there is an element which masks all DOM
element (it is located on top of all the element and fills the entire
body element)
As a result when clicking a control it does not receive the event -> the
mask element will receive the event.
In this change we delegate the event to the control behind the mask.