When adding n values to multi, avoiding looping over them n times
Currently, every time you call $input.select2('val', array), the
setVal() function is called for every array entry, and setVal() loops
over the entire array of values to ensure uniqueness. When the number
of values being set is fairly large (on the order of hundreds), this has
a dramatic impact on rendering performance.

This commit adds a new internal method, setUniqueVal(), to call when a
value array is already known to be unique.