Add placeholder support to angular models
When the ngModel directive is used, angular automatically inserts an
empty <option value="?" selected="selected"></option> element if the
model does not have any value selected. This prevents the correct
functionality of the placeholder because the value echoed by angular is
not empty (rather, it prints a question mark). The only alternative
would be passing a custom function to every select2 instance for
comparison.