Switch remote example to GitHub Search API
The previous Rotten Tomatoes API had a hard usage limit per day
that was tied to the API key.  It was very common for the usage
limit to be hit by the end of the day, which caused a bad
experience for those interested in trying out Select2 with a remote
API.

The GitHub Search API was chosen as the replacement because it
allows for anonymous usage, so we do not need to tie anything to
tokens, and it is rate-limited to the minute instead of the day.
These two things combined together allows users to use the example
without having to worry as much about it stopping for the rest of
the day.

The formatters and examples were swapped out to use the API, and
a custom formatter was created to replace the old one.