Added the "emptySearchText" option
Added an option which allows the developer to specify a text which is diplayed under the search box when tehre is no input. The original behaviour was that in this case select2 shows the "too short" message, which is not too informative here. With this option the user interface can be improved a lot. 

Further improvement could be to give it some default, and make it localizable.